Relocate auth forms components
