+++ app/(auth)/token-error/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useSearchParams } from 'next/navigation';
 
 export default function Page() {
